Fix some MSVC compilation errors

+++ mlx/backend/cuda/reduce/all_reduce.cu
@@ -68,8 +68,8 @@ void all_reduce(
 
   out.set_data(cu::malloc_async(out.nbytes(), encoder));
 
-  auto get_args = [](size_t size, int N) {
-    int threads = std::min(512UL, (size + N - 1) / N);
+  auto get_args = [](int size, int N) {
+    int threads = std::min(512, (size + N - 1) / N);
     threads = ((threads + WARP_SIZE - 1) / WARP_SIZE) * WARP_SIZE;
     int reductions_per_step = threads * N;
     size_t steps_needed =
